Refactor user registration functionality and structure backend code
