+++ .gitignore
@@ -5,7 +5,7 @@
 /.pnp
 .pnp.js
 /backend/node_modules
-/backend/dist/*.*
+/backend/dist/**
 
 # testing
 /coverage
