+++ backend/src/routes/userRoutes.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import UserController from '../controllers/userController';
+
+const userRoutes = Router();
+
+userRoutes.post('/register', UserController.register);
+
+export default userRoutes;
